+++ .storybook/main.js
@@ -1,16 +1,18 @@
 module.exports = {
   stories: [
     // files MUST end with .story.mdx or .stories.mdx (@see https://github.com/storybookjs/storybook/issues/9918)
-    '../src/components/**/*.story.mdx',
-    '../src/components/**/*.stories.mdx',
+    "../src/components/**/*.story.mdx",
+    "../src/components/**/*.stories.mdx",
+    "../src/snowflakes/**/*.story.mdx",
+    "../src/snowflakes/**/*.stories.mdx",
   ],
   addons: [
-    '@storybook/addon-viewport/register',
-    '@storybook/addon-knobs/register',
-    '@storybook/addon-a11y/register',
-    'storybook-addons-abstract',
+    "@storybook/addon-viewport/register",
+    "@storybook/addon-knobs/register",
+    "@storybook/addon-a11y/register",
+    "storybook-addons-abstract",
     {
-      name: '@storybook/addon-docs',
+      name: "@storybook/addon-docs",
       options: {
         configureJSX: true,
         babelOptions: {},