+++ src/components/modal/z-modal-login/index.spec.ts
@@ -132,8 +132,8 @@ const renderUsername = (hidden = false, status?, message?) =>
   `<slot-fb name="username" ${hidden ? `hidden` : ``}>
     <z-input
       id="username"
-      label="email o numero di telefono"
-      placeholder="Inserisci la mail o il cellulare"
+      label="email o numero di cellulare"
+      placeholder="Inserisci l'email o il cellulare"
       autocomplete="username"
       name="username"
       ${status ? `status="${status}"` : ``}
@@ -146,7 +146,7 @@ const renderPassword = (hidden = false, status?) =>
     <z-input
       id="password"
       label="password"
-      placeholder="Inserisci la password"
+      placeholder="Inserisci la tua password"
       type="password"
       name="password"
       autocomplete="current-password"
@@ -161,7 +161,7 @@ const renderForgottenPwd = () =>
 
 const renderLogin = (hidden = false) =>
   `<slot-fb name="login" ${hidden ? `hidden` : ``}>
-    <z-button variant="primary">Login</z-button>
+    <z-button variant="primary">Accedi</z-button>
   </slot-fb>`;
 
 const renderSignupBody = () =>